requirements.txt
# ============================================
# AI Document Intelligence System v2.0
# Stable Requirements (Python 3.10+)
# ============================================

# Core Framework
streamlit==1.38.0
python-dotenv==1.0.1

# ============================================
# AI & Machine Learning
# ============================================
openai==1.45.0
tiktoken==0.7.0

# LangChain Ecosystem (Stable)
langchain==0.3.21
langchain-core==0.3.49
langchain-community==0.3.1
langchain-openai==0.2.0

# Embeddings & Transformers
torch==2.3.0
transformers==4.39.3
sentence-transformers==3.0.1

# Speech Recognition
openai-whisper==20231117
SpeechRecognition==3.10.0

# ============================================
# Database & Storage
# ============================================
chromadb==0.5.11
sqlite-utils==3.35

# ============================================
# Document Processing
# ============================================
pdfplumber==0.10.3
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.3.0                  # Pillow>=10.0.0
opencv-python==4.10.0.84        # opencv-python>=4.8.0

# Excel & Data
pandas==2.2.2
openpyxl==3.1.2
xlrd==2.0.1

# ============================================
# TEXT-TO-SPEECH & AUDIO/VIDEO GENERATION
# ============================================
moviepy==1.0.3
gtts==2.4.0
deep-translator==1.11.4
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
imageio-ffmpeg==0.4.9       

# ============================================
# Visualization & Analytics
# ============================================
plotly==5.24.1
matplotlib==3.9.2
seaborn==0.13.2

# ============================================
# Data Science & Analysis
# ============================================
numpy==1.26.4           #numpy>=1.24.0
scipy==1.13.1
scikit-learn==1.5.2

# ============================================
# Web & HTTP
# ============================================
requests==2.32.3
httpx==0.27.2
aiohttp==3.10.8

# ============================================
# Translation & Language
# ============================================
googletrans==4.0.2
deep-translator==1.11.4
dateparser==1.2.0

# ============================================
# Utilities
# ============================================
pydantic==2.9.2
pydantic-settings==2.6.1
typing-extensions==4.12.2
tqdm==4.66.5
pathlib==1.0.1
hashlib-additional==1.0.0

# ============================================
# Logging & Monitoring
# ============================================
loguru==0.7.2
colorlog==6.8.2

# ============================================
# Testing & Development
# ============================================
pytest==8.3.3
pytest-cov==5.0.0
pytest-asyncio==0.23.7
black==24.10.0
flake8==7.1.1
mypy==1.11.2
pre-commit==3.8.0

# ============================================
# Production & Deployment
# ============================================
gunicorn==23.0.0
uvicorn[standard]==0.31.0
python-multipart==0.0.9
prometheus-client==0.21.0
psutil==6.0.0

# API Server
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.9


# ============================================
# Optional: Cloud & DB
# ============================================
# boto3==1.34.162
# botocore==1.34.162
# pymongo==4.8.0
# redis==5.0.8

# ============================================
# Platform Specific
# ============================================
pywin32==306  # Windows
# python-magic==0.4.27  # Linux/macOS

# ============================================
# Certificates & Networking
# ============================================
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
urllib3==2.2.3
# ============================================
# Metrics & Analytics (ADD THESE)
# ============================================
xlsxwriter==3.2.0

# pip install python-docx python-pptx mammoth beautifulsoup4 markdown striprtf ebooklib
# pip install hf_xet